Backend/requirements.txt
groq
playwright
python-dotenv
pandas
sqlalchemy 
psycopg2-binary
python-telegram-bot